Updated Functions class names and moved methods to their respective classes
