+++ src/main/java/com/metrics/service/MetricsServiceImpl.java
@@ -89,7 +89,7 @@ public class MetricsServiceImpl implements MetricsService {
 		MetricsCollection metric = new MetricsCollection();
 		MetricsApplication.logger.info("calling data validation method");
 		metric = TechnicalValidations
-				.CreateMetricRequestToMetricsCollection(ClientValidations.testMetricIntegrity(request, 1));
+				.CreateMetricRequestToMetricsCollection(BusinessMethods.testMetricIntegrity(request, 1));
 		MetricsApplication.logger.info("data validation test passed, saving new object");
 		metric = orikaMapperFacade.map(request, MetricsCollection.class);
 		MetricsApplication.logger.info("object created succesfully");
